Bug 1102208 - It takes long time to turn on FM Radio in Flame
msm8610 doesn't support analog audio routing. If we try to enable it
anyway, a 4 seconds delay is introduced when starting the fm radio, due
to the wait for hw.fm.init to be set in gecko
(hal/gonk/GonkFMRadio.cpp). By setting ro.moz.fm.noAnalog to true, we
won't try to enable it.